# app/docs/swagger.yaml
definitions:
  models.AccessTokenOut:
    properties:
      token:
        type: string
    type: object
  models.OpeningOut:
    properties:
      anime_name:
        type: string
      codename:
        type: string
      id:
        type: integer
      name:
        type: string
      thumbnail_link:
        type: string
      youtube_embed_link:
        type: string
    type: object
  models.UserOut:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: Authorize user
      parameters:
      - description: Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccessTokenOut'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Authorize user
      tags:
      - auth
  /openings:
    get:
      consumes:
      - application/json
      description: Get list of openings by limit and offset
      parameters:
      - description: Count of resulting rows
        in: query
        name: limit
        type: integer
      - description: Offset of the query
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OpeningOut'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get list of openings by limit and offset
      tags:
      - openings
  /openings/{codename}:
    get:
      consumes:
      - application/json
      description: Get one opening by its codename
      parameters:
      - description: Codename of opening to get
        in: path
        name: codename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OpeningOut'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get one opening by its codename
      tags:
      - openings
  /profile/{username}:
    get:
      consumes:
      - application/json
      description: Get profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserOut'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get profile by username
      tags:
      - profile
  /register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AccessTokenOut'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Register user
      tags:
      - auth
swagger: "2.0"